src/core/api/permissions.py
from rest_framework.permissions import BasePermission


class IsAdminRole(BasePermission):
    """Allows access only to users with the admin role."""

    def has_permission(self, request, view):
        return (
            request.user
            and request.user.is_authenticated
            and getattr(request.user, "role", None) == "admin"
        )
