requirements/base.txt
Django==5.1.4
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.28.0
